backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
bcrypt==3.2.0
PyJWT==2.3.0
python-decouple==3.5
dnspython==2.1.0 